Add files via upload
   Level One, Task One-  A Landing Page:
Here is a Landing Page for a website encouraging learners to speak English fluently.  I have created this with the help of basic knowledge of HTML and CSS and learnt to create columns, divide sections ,arrange items and add headers and footers. I have done this task during my internship in web development@ CodSoft.  
#CodSoft
#Web Development